Release v0.9.6

+++ src/pages/Scheduled.jsx
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react';
 import { supabase } from '../supabaseClient';
-import Dashboard from '../components/Dashboard';
 import Notifications from '../components/Notifications';
 import ScheduledList from '../components/ScheduledList';
 import PropTypes from 'prop-types';
@@ -37,29 +36,36 @@ const Scheduled = ({ session }) => {
     // allow user to delete item from list
 
     const handleDelete = async (flight) => {
-        const { error } = await supabase
-            .from('schedule_lookup')
-            .delete()
-            .eq('id', flight.id);
-
-        if (error) {
-            console.warn(error);
-        }
-        setScheduledFlights(
-            scheduledFlights.filter((item) => item.id !== flight.id)
+        const confirm = window.confirm(
+            `Are you sure you want to delete ${flight.ident}?`
         );
+        if (confirm) {
+            const { error } = await supabase
+                .from('schedule_lookup')
+                .delete()
+                .eq('id', flight.id);
+
+            if (error) {
+                console.warn(error);
+            }
+            setScheduledFlights(
+                scheduledFlights.filter((item) => item.id !== flight.id)
+            );
+        }
     };
 
     return (
         <div>
-            <Dashboard key={session.user.id} session={session} />
             <Notifications />
-            <div className='px-4 py-5 sm:px-6 max-w-md'>
+            <div className='px-4 py-5 sm:px-6 max-w-lg'>
                 <h3 className='text-base font-semibold leading-6 text-white'>
                     Scheduled
                 </h3>
                 <p className='mt-1 max-w-2xl text-sm text-gray-500'>
-                    These are searches that are scheduled.
+                    These are searches that are scheduled.{' '}
+                    {scheduledFlights.length > 0
+                        ? `You have ${scheduledFlights.length} scheduled searches.`
+                        : 'You have no scheduled searches.'}
                 </p>
             </div>
             <div className='border-t border-zinc-600'>
